+++ modules/@apostrophecms/doc-type/ui/apos/logic/AposDocContextMenu.js
@@ -405,10 +405,10 @@ export default {
         this.menuHandler('discardDraft');
       }
     },
-    async onContentChanged(e) {
-      if (e.doc && (e.doc._id === this.context._id)) {
-        this.context = e.doc;
-      } else if (e.docIds && e.docIds.includes(this.context._id)) {
+    async onContentChanged({ doc, docIds }) {
+      if (doc && (doc._id === this.context._id)) {
+        this.context = doc;
+      } else if (docIds && docIds.includes(this.context._id)) {
         try {
           this.context = await apos.http.get(`${this.moduleOptions.action}/${this.context._id}`, {
             busy: true