+++ modules/@apostrophecms/image/ui/apos/components/AposMediaManager.vue
@@ -70,9 +70,11 @@
             :checked-count="checked.length"
             :module-name="moduleName"
             :options="{noPager: true}"
+            :batch-operations="moduleOptions.batchOperations"
             @select-click="selectClick"
             @search="search"
             @filter="filter"
+            @batch="handleBatchAction"
           />
         </template>
         <template #bodyMain>
@@ -249,7 +251,9 @@ export default {
       }
     },
     isLastPage() {
-      return this.totalPages > 1 && this.currentPage === this.totalPages;
+      return this.totalPages > 1 &&
+        this.currentPage === this.totalPages &&
+        !this.isScrollLoading;
     }
   },
 
@@ -412,7 +416,7 @@ export default {
       await this.debouncedGetMedia.skipDelay(opts);
       this.isLoading = false;
       this.modified = false;
-      this.updateEditing(null);
+      this.clearSelected();
     },
     async filter(name, value) {
       this.filterValues[name] = value;
@@ -552,20 +556,19 @@ export default {
       this.isLoading = false;
       this.skipLoadObserver = false;
     },
-
-    async onContentChanged({ action, doc }) {
-      if (doc.type !== '@apostrophecms/image' || ![ 'archive', 'update' ].includes(action)) {
+    async onContentChanged({
+      action, doc, docTypes
+    }) {
+      const types = this.getContentChangedTypes(doc, docTypes);
+      if (!types.includes(this.moduleName)) {
         return;
       }
-
       this.modified = false;
-      if (action === 'archive') {
-        this.removeStateDoc(doc);
-      }
       if (action === 'update') {
         this.updateStateDoc(doc);
+      } else {
+        this.refetchMedia();
       }
-
       await this.updateEditing(null);
     },
 
@@ -581,6 +584,8 @@ export default {
       }
     },
 
+    // Keep it for later when we will be able to udpate the UI without refreshing existing
+    // because it would break pagination.
     removeStateDoc(doc) {
       const index = this.items.findIndex(item => item._id === doc._id);
       const checkedIndex = this.checked.findIndex(checkedId => checkedId === doc._id);
@@ -664,6 +669,29 @@ export default {
       this.disconnectObserver();
       if (ref) {
         this.observeLoadRef();
+      }
+    },
+    async handleBatchAction({
+      label, action, requestOptions = {}, messages
+    }) {
+      if (action) {
+        try {
+          await apos.http.post(`${this.moduleOptions.action}/${action}`, {
+            body: {
+              ...requestOptions,
+              _ids: this.checked,
+              messages,
+              type: this.checked.length === 1 ? this.moduleLabels.singular
+                : this.moduleLabels.plural
+            }
+          });
+        } catch (error) {
+          apos.notify('apostrophe:errorBatchOperationNoti', {
+            interpolate: { operation: this.$t(label) },
+            type: 'danger'
+          });
+          console.error(error);
+        }
       }
     }
   }