+++ modules/@apostrophecms/image/ui/apos/components/AposMediaManagerEditor.vue
@@ -72,7 +72,7 @@
           <AposButton
             type="quiet"
             label="apostrophe:download"
-            :href="!isArchived ? activeMedia.attachment._urls.original : false"
+            :href="!isArchived ? activeMedia.attachment._urls.original : null"
             :disabled="isArchived"
             download
           />
@@ -340,7 +340,7 @@ export default {
         });
         apos.bus.$emit('content-changed', {
           doc,
-          action: 'update'
+          action: this.restoreOnly ? 'restore' : 'update'
         });
         this.original = klona(this.docFields.data);
       } catch (e) {