+++ modules/@apostrophecms/job/index.js
@@ -121,7 +121,8 @@ module.exports = {
             // can't show progress.
             jobId: total && job._id,
             ids,
-            action: options.action
+            action: options.action,
+            docTypes: options.docTypes
           });
 
           return {
@@ -197,7 +198,10 @@ module.exports = {
           job = await self.start(options);
 
           const notification = await self.triggerNotification(req, 'progress', {
-            jobId: job._id
+            jobId: job._id,
+            ids: options.ids,
+            action: options.action,
+            docTypes: options.docTypes
           });
 
           run({ notificationId: notification.noteId });
@@ -284,7 +288,8 @@ module.exports = {
           job: {
             _id: options.jobId,
             action: options.action,
-            ids: options.ids
+            ids: options.ids,
+            docTypes: options.docTypes
           },
           event,
           classes: options.classes,