+++ modules/@apostrophecms/ui/ui/apos/components/AposTree.vue
@@ -27,7 +27,9 @@
       :options="options"
       :tree-id="treeId"
       :module-options="moduleOptions"
+      :expanded-index="expandedIndex"
       @update="update"
+      @toggle="onToggleSection"
     />
   </div>
 </template>
@@ -87,7 +89,8 @@ export default {
       myItems: klona(this.items),
       nested: false,
       colWidths: null,
-      treeId: createId()
+      treeId: createId(),
+      expandedIndex: {}
     };
   },
   computed: {
@@ -196,6 +199,9 @@ export default {
         // The index of the moved item within its new context.
         endIndex: event.newIndex
       });
+    },
+    onToggleSection({ _id, expanded }) {
+      this.expandedIndex[_id] = expanded;
     }
   }
 };